+++ src/check.rs
@@ -253,18 +253,18 @@ impl<'a> VirtualEnv<'a> {
                         }
                     }
                 }
-                Bind => {
+                Do => {
                     let f = self.pop_func()?;
                     let g = self.pop_func()?;
-                    for val in [g, f] {
-                        for _ in 0..val.signature().args {
-                            self.pop()?;
-                        }
-                        self.set_min_height();
-                        for _ in 0..val.signature().outputs {
-                            self.stack.push(BasicValue::Other);
-                        }
-                    }
+                    let f_sig = f.signature();
+                    let g_sig = g.signature();
+                    let copy_count = g_sig.args.saturating_sub(g_sig.outputs - 1);
+                    let g_sub_sig = Signature::new(g_sig.args, g_sig.outputs + copy_count - 1);
+                    let comp_sig = f_sig.compose(g_sub_sig);
+                    self.handle_args_outputs(
+                        comp_sig.args,
+                        comp_sig.outputs + g_sub_sig.outputs.saturating_sub(g_sig.args),
+                    )?;
                 }
                 Both => {
                     let sig = self.pop_func()?.signature();