+++ changelog.md
@@ -5,10 +5,13 @@ Uiua is not yet stable.
 ## 0.0.26 - 2023-11-??
 If you are reading this on the website, then these changes are live here.
 ### Language
+- Add the [`do` `⍢`](https://uiua.org/docs/do) modifier, which repeatedly calls a function as long as a condition holds
+- Deprecate [`break` `⎋`](https://uiua.org/docs/break)
 - Add multi-dimensional [`where` `⊚`](https://uiua.org/docs/where)
 - [`join` `⊂`](https://uiua.org/docs/join) to an empty list now always works regardless of rank of the other array
 - [`each` `∵`](https://uiua.org/docs/each) and [`rows` `≡`](https://uiua.org/docs/rows) now work with [`under` `⍜`](https://uiua.org/docs/under)
 - Parsing multiple formattable functions from words is now smarter
+- Remove `bind '`. It made code hard to read. It will continue to parse, but will be formatted as `(…)`
 ### Interpreter
 - Add the `uiua stand` command, which creates a standalone executable